Rename package commandresult to results
