+++ src/main/java/seedu/mark/logic/commands/Command.java
@@ -1,7 +1,7 @@
 package seedu.mark.logic.commands;
 
-import seedu.mark.logic.commands.commandresult.CommandResult;
 import seedu.mark.logic.commands.exceptions.CommandException;
+import seedu.mark.logic.commands.results.CommandResult;
 import seedu.mark.model.Model;
 
 /**
