+++ src/main/java/seedu/mark/logic/commands/ExitCommand.java
@@ -1,7 +1,7 @@
 package seedu.mark.logic.commands;
 
-import seedu.mark.logic.commands.commandresult.CommandResult;
-import seedu.mark.logic.commands.commandresult.ExitCommandResult;
+import seedu.mark.logic.commands.results.CommandResult;
+import seedu.mark.logic.commands.results.ExitCommandResult;
 import seedu.mark.model.Model;
 
 /**
