+++ src/main/java/seedu/mark/logic/commands/ListCommand.java
@@ -3,7 +3,7 @@ package seedu.mark.logic.commands;
 import static java.util.Objects.requireNonNull;
 import static seedu.mark.model.Model.PREDICATE_SHOW_ALL_BOOKMARKS;
 
-import seedu.mark.logic.commands.commandresult.CommandResult;
+import seedu.mark.logic.commands.results.CommandResult;
 import seedu.mark.model.Model;
 
 /**
