+++ src/main/java/seedu/mark/logic/commands/results/ExitCommandResult.java
@@ -1,4 +1,4 @@
-package seedu.mark.logic.commands.commandresult;
+package seedu.mark.logic.commands.results;
 
 /**
  * Represents the result of an exit command execution.
