+++ src/test/java/seedu/mark/logic/commands/ExitCommandTest.java
@@ -5,8 +5,8 @@ import static seedu.mark.logic.commands.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEMENT
 
 import org.junit.jupiter.api.Test;
 
-import seedu.mark.logic.commands.commandresult.CommandResult;
-import seedu.mark.logic.commands.commandresult.ExitCommandResult;
+import seedu.mark.logic.commands.results.CommandResult;
+import seedu.mark.logic.commands.results.ExitCommandResult;
 import seedu.mark.model.Model;
 import seedu.mark.model.ModelManager;
