+++ src/test/java/seedu/mark/logic/commands/HelpCommandTest.java
@@ -5,8 +5,8 @@ import static seedu.mark.logic.commands.HelpCommand.SHOWING_HELP_MESSAGE;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.mark.logic.commands.commandresult.CommandResult;
-import seedu.mark.logic.commands.commandresult.HelpCommandResult;
+import seedu.mark.logic.commands.results.CommandResult;
+import seedu.mark.logic.commands.results.HelpCommandResult;
 import seedu.mark.model.Model;
 import seedu.mark.model.ModelManager;
